# [WebGPU] Direct CPU->GPU buffer upload for UMA