--- a/onnxruntime/core/providers/webgpu/webgpu_context.h
+++ b/onnxruntime/core/providers/webgpu/webgpu_context.h
@@ -143,6 +143,8 @@ class WebGpuContext final {
   Status Run(ComputeContext& context, const ProgramBase& program);
   void OnRunEnd();
 
+  bool SupportsBufferMapExtendedUsages() const { return supports_buffer_map_extended_usages_; }
+
  private:
   enum class TimestampQueryType {
     None = 0,
@@ -231,6 +233,8 @@ class WebGpuContext final {
 #if defined(ENABLE_PIX_FOR_WEBGPU_EP)
   std::unique_ptr<WebGpuPIXFrameGenerator> pix_frame_generator_ = nullptr;
 #endif  // ENABLE_PIX_FOR_WEBGPU_EP
+
+  bool supports_buffer_map_extended_usages_ = false;
 };
 
 }  // namespace webgpu