This eliminates the using of staging buffers when uploading data from CPU to GPU for UMA. The data is directly copied to the GPU buffer. The major challenge of enabling this feature is to avoid waiting for the mapping of the dest GPU buffer. The idea is to map the dest GPU buffers in advance, so that when the uploading starts, the dest GPU is always mapped. This is achieved by:
1. using map-at-creation for all storage buffers.
2. mapping the buffer once it's returned, and marking it as ready for reuse after the mapping is done.
3. checking to unmap the buffer whereever it's used by the command encoder.