@fs-eire 2025-03-07
@jchen10 thank you for the optimization!

I checked the [document](https://dawn.googlesource.com/dawn/+/refs/heads/main/docs/dawn/features/buffer_map_extended_usages.md) of the feature. As it described, this feature allows to map the data on the storage buffer and makes it possible to get rid of the staging buffer for uploading/downloading.

For downloading it looks like this feature is a simple optimization. However for uploading the situation became a little bit complicated.

As already mentioned in the description of this PR, to enable this feature for uploading, we need the buffer to be in "mapped" state. If we call `buffer.mapAsync` in buffer upload, it will cause a GPU-CPU synchronization, which breaks the pipeline and may cause perf problem. However, creating all storage buffer with mapped=true may also cause other problems:
- Some buffers will never map to CPU in their whole lifecycle. Creating them as mapped and unmap them before use is unnecessary and will cause more memory usage (whether this impact performance or not need to run test and figure out by data)
- For some usage (for example, in wasm32), upload buffers are not mapped inside C++. Instead, the code opens holes using `EM_ASM` to directly upload data from outside of the wasm. Without this trick, many models may run out of 4GB memory and fail the wasm runtime.


I think the real problem is that the current implementation of ORT does not give the context information to allocator so `BufferManager` has no idea what is the usage of a buffer that it is creating. With this limitation, we have to choose either creating all buffers mapped or all buffers unmapped. Ideally ORT should give some context info so that buffer manager would do better job.

There may need more discussion about how to use this feature in ORT.