@jchen10 2025-03-07
@fs-eire Thanks for the comment. Absolutely it would be perfect if ORT could give more context info to the buffer manager. Let's investigate this further.
BTW, this optimization looks promising. On my TGL, it shows 5.1G->3.2G memory reduction without performance regression.
Without UMA:
```
model_benchmark.exe -i ..\models-genai\Phi-3.5-mini-instruct-onnx-web
Batch size: 1, prompt tokens: 17, tokens to generate: 128
Prompt processing (time to first token):
        avg (us):       286253
        avg (tokens/s): 59.388
        p50 (us):       294972
        stddev (us):    12417.6
        n:              5 * 17 token(s)
Token generation:
        avg (us):       56041.5
        avg (tokens/s): 17.8439
        p50 (us):       56055.8
        stddev (us):    1682.57
        n:              635 * 1 token(s)
Token sampling:
        avg (us):       15.64
        avg (tokens/s): 63938.6
        p50 (us):       15.8
        stddev (us):    0.364692
        n:              5 * 1 token(s)
E2E generation (entire generation loop):
        avg (ms):       7403.63
        p50 (ms):       7393.68
        stddev (ms):    20.0764
        n:              5
Peak working set size (bytes): 5530017792
```
With UMA:
```
model_benchmark.exe -i ..\models-genai\Phi-3.5-mini-instruct-onnx-web
Batch size: 1, prompt tokens: 17, tokens to generate: 128
Prompt processing (time to first token):
        avg (us):       275511
        avg (tokens/s): 61.7036
        p50 (us):       273323
        stddev (us):    6202.55
        n:              5 * 17 token(s)
Token generation:
        avg (us):       56323.8
        avg (tokens/s): 17.7545
        p50 (us):       56297.5
        stddev (us):    1842.57
        n:              635 * 1 token(s)
Token sampling:
        avg (us):       12
        avg (tokens/s): 83333.3
        p50 (us):       14.9
        stddev (us):    4.01933
        n:              5 * 1 token(s)
E2E generation (entire generation loop):
        avg (ms):       7428.74
        p50 (ms):       7417.44
        stddev (ms):    24.1045
        n:              5
Peak working set size (bytes): 3419979776
```